Renamed IncomingRequest to ServerRequest
Per the change in PSR-7 interfaces, renamed the incoming request and
related factory to "server request". Ensured that all methods exposed in
the interface are now present.

Cleaned up the `fromGlobals()` method to remove transient variables.

ImmutableHeadersTrait is no longer necessary, so removed.
